Redis operator (#1)
This PR adds the initial version of the Redis Operator for K8s. It was based on the mongodb (https://github.com/canonical/mongodb-operator), elasticsearch(https://github.com/canonical/elasticsearch-operator/) and this (https://github.com/canonical/charm-mssql/pull/6/files) PR.

It was made with the goal to first serve for the Commercial Systems team to deploy a Redis cache without replication. Therefore, at first it does not provide Redis cluster capabilities.